* W32: Fixes the last patch (trim) to work in the microsoft compiler as  	well.
git-svn-id: svn://svn.minisip.org/minisip/trunk@1745 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ libmutil/source/trim.cxx
@@ -32,6 +32,7 @@
 */
 
 #include<libmutil/trim.h>
+#include<libmutil/mtypes.h>
 #include<ctype.h>
 
 #include<string>
@@ -43,12 +44,12 @@ LIBMUTIL_API string trim(string line){
 	int32_t idx;
 
 	idx = 0;
-	while( idx < line.size() && isspace(line[idx]) ) {
+	while( idx < (int)line.size() && isspace(line[idx]) ) {
 		spacesFront++;
 		idx++;
 	}
 	
-	idx = line.size() - 1 ;
+	idx = (int)line.size() - 1 ;
 	while( idx >= 0 && isspace(line[idx]) ) {
 		spacesEnd++;
 		idx--;
